requirements.txt
pyyaml
flask
pip